+++ dsh_sdk/src/schema_store/types/subject_strategy.rs
@@ -127,8 +127,6 @@ impl Hash for SubjectName {
 
 #[cfg(test)]
 mod tests {
-    use openssl::hash;
-
     use super::*;
     use std::hash::DefaultHasher;