+++ dsh_sdk/src/utils/mod.rs
@@ -12,7 +12,7 @@ use crate::error::DshError;
 pub mod dlq;
 #[cfg(feature = "graceful-shutdown")]
 pub mod graceful_shutdown;
-#[cfg(feature = "hyper-client")]
+#[cfg(feature = "hyper-client")] // TODO: to be implemented
 pub(crate) mod http_client;
 #[cfg(feature = "metrics")]
 pub mod metrics;